Change for simulation without space and distance consideration.
