+++ population.go
@@ -5,8 +5,7 @@ import (
 	"math/rand"
 	"time"
 
-	domain "github.com/arbori/population.git/population/domain"
-	simulation "github.com/arbori/population.git/population/usecase"
+	simulation "github.com/arbori/population.git/population/simulation"
 )
 
 func main() {
@@ -14,33 +13,37 @@ func main() {
 
 	fmt.Println("Simulated society")
 
-	simulationData := domain.SimulationData{
+	society := simulation.SocietyParameters{
 		IndividualConsumption: 2000,
 		IndividualExchange:    2000,
 		SocietySize:           50000,
 		ViabilityAmount:       0,
 		SolidarityProbability: 0.0,
-		Ephocs:                0,
 	}
 
-	viabilityRange := domain.Range{
+	simulationData := simulation.SocietyRealization{
+		Society: society,
+		Ephocs:  0,
+	}
+
+	viabilityRange := simulation.Range{
 		Minimum: 500,
 		Maximum: 10500,
 		Delta:   500,
 	}
 
-	solidarityRange := domain.Range{
+	solidarityRange := simulation.Range{
 		Minimum: float32(0.0),
 		Maximum: float32(1.0),
 		Delta:   float32(0.1),
 	}
 
-	simulationSurface := simulation.Run(simulationData, viabilityRange, solidarityRange)
+	simulationSurface := simulationData.Run(viabilityRange, solidarityRange)
 
-	fmt.Printf("IndividualConsumption: %d\nIndividualExchange: %d\nSocietySize: %d\n", simulationData.IndividualConsumption, simulationData.IndividualExchange, simulationData.SocietySize)
+	fmt.Printf("IndividualConsumption: %d\nIndividualExchange: %d\nSocietySize: %d\n", simulationData.Society.IndividualConsumption, simulationData.Society.IndividualExchange, simulationData.Society.SocietySize)
 	fmt.Println("ViabilityAmount\tSolidarityProbability\tEphocs")
 
 	for i, data := range simulationSurface {
-		fmt.Printf("%d\t%d\t%.2f\t%d\n", i, data.ViabilityAmount, data.SolidarityProbability, data.Ephocs)
+		fmt.Printf("%d\t%d\t%.2f\t%d\n", i, data.Society.ViabilityAmount, data.Society.SolidarityProbability, data.Ephocs)
 	}
 }
